% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAndFormatXlsx.R
\name{createAndFormatXlsx}
\alias{createAndFormatXlsx}
\title{Uses the output from water_balance_by_station and creates formats and writes an excell file .xlsx}
\usage{
createAndFormatXlsx(df, filepath, trgtM = Sys.Date())
}
\arguments{
\item{df}{a dataframe as output form water_balance_by_station}

\item{filepath}{the absolute path of the xlsx to write. has to include extension .xlsx}

\item{trgtM}{a date from wich the function derives the abbreviation of the name in german. Used to name the excell sheet}
}
\description{
Uses the output from water_balance_by_station and creates formats and writes an excell file .xlsx
}
