% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ET}
\alias{ET}
\title{Computes Evapotraspiration using PenmanMonteith from data retrieved with function mergeData}
\usage{
ET(data, Kc.ini.A = 0.1, Kc.ini.B = 0.4, Kc.dev = 0.7,
  Kc.mid = 1.2, Kc.late = 0.8, DOY.ini.A = 69, DOY.ini.B = 71,
  DOY.dev = 91, DOY.mid = 141, DOY.late = 271, DOY.harv = 312,
  tree_height = 3, slopeCorr = NULL, radiation_unit = "W",
  crop = "short", netRadiation = "no", wind = "yes",
  message = "yes")
}
\description{
Computes Evapotraspiration using PenmanMonteith from data retrieved with function mergeData
}
