% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_PenmanMonteith.R
\name{ET_PenmanMonteith}
\alias{ET_PenmanMonteith}
\title{Computes Evapotraspiration using PenmanMonteith: modified from evapotranspiration package
in order to choose between net or global radiation}
\usage{
ET_PenmanMonteith(data, constants, ts = "daily",
  solar = "sunshine hours", wind = "yes", crop = "short",
  message = "yes", save.csv = "yes", netRadiation = "no", ...)
}
\description{
Computes Evapotraspiration using PenmanMonteith: modified from evapotranspiration package
in order to choose between net or global radiation
}
