% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BR_data.R
\name{get_BR_data}
\alias{get_BR_data}
\title{Get data from SBR monitoring stations.}
\usage{
get_BR_data(station, datestart = Sys.Date() - 1, dateend = Sys.Date() +
  1, user, password, host, add_names = FALSE, spread = FALSE,
  round = "hour")
}
\arguments{
\item{station}{station number.}

\item{datestart}{starting date of the timeseries.}

\item{dateend}{ending date of the timeseries.}
}
\description{
Get data from SBR monitoring stations.
}
