% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosestStations.R
\name{getClosestStations}
\alias{getClosestStations}
\title{Get closest province of Bozen and SBR stations}
\usage{
getClosestStations(long, lat, provSensor = NULL)
}
\description{
Get closest province of Bozen and SBR stations
}
