% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram.R
\name{gram}
\alias{gram}
\title{Function gram}
\usage{
gram(X, trX, block = FALSE, block.size = 1000, show.progress = FALSE)
}
\arguments{
\item{X}{a standardized feature matrix.}

\item{trX}{(optional) the transpose matrix of X.}

\item{block}{logical, block matrix computation is performed when TRUE, default option is FALSE.}

\item{block.size}{used when block = TRUE. Default option is 1000, i.e. blocks of dimensionality 1000 x 1000.}

\item{show.progress}{logical, when TRUE (and block = TRUE) the progress of the calculations is reported on the console.}
}
\value{
Returns the inner-product Gram matrix.
}
\description{
Function for calculating the (inner-product) Gram matrix that allows for block-matrix multiplication.
}
\examples{
X <- matrix(rnorm(100 * 300), 100, 300)
G0 <- gram(X)                             # usual matrix multiplication                          
G1 <- gram(X, block=TRUE,block.size=100)  # block matrix multiplication
all.equal(G0, G1)  
}
\references{
Perrakis, K., Mukherjee, S and the Alzheimers Disease Neuroimaging Initiative. (2018) Scalable Bayesian regression in high dimensions with multiple data sources. \url{https://arxiv.org/pdf/1710.00596.pdf}
}
\seealso{
\code{\link{gram.parallel}}
}
\author{
Konstanstinos Perrakis \email{konstantinos.perrakis@dzne.de}

Sach Mukherjee \email{sach.mukherjee@dzne.de}
}
