\name{hh}
\alias{hh}
\docType{data}
\title{Header for Humans (hh)}
\description{
  Information about human genes in GSE147507.  
}
\usage{data(hh)}
\format{
  A data frame with the following columns. Only rows with non-blank Symbols are included.
  \describe{
    \item{\code{Symbol}}{Gene symbol (short acronym). }
    \item{\code{Length}}{Length of gene. Used by tpm().}
  }
}



\keyword{datasets}
