% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeEtAndForecast.R
\name{mergeEtAndForecast}
\alias{mergeEtAndForecast}
\title{merge data from wheater forecast and et average data}
\usage{
mergeEtAndForecast(long, lat, slope = NULL, Kc.ini.A = 0.1,
  Kc.ini.B = 0.4, Kc.dev = 0.7, Kc.mid = 1.2, Kc.late = 0.8,
  DOY.ini.A = 69, DOY.ini.B = 71, DOY.dev = 91, DOY.mid = 141,
  DOY.late = 271, DOY.harv = 312)
}
\description{
merge data from wheater forecast and et average data
}
