% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOldAndForecast.R
\name{mergeOldAndForecast}
\alias{mergeOldAndForecast}
\title{merge et and precipitation data from wheater forecast and dayly measured et data}
\usage{
mergeOldAndForecast(data, long, lat, taw = 50, lmitWarning = 0.7,
  p = 0.5, start = 0)
}
\description{
merge et and precipitation data from wheater forecast and dayly measured et data
}
