\name{pDf}
\alias{pDf}
\docType{data}
\title{Pheno Data Ferret (pDf)}
\description{
  Information about the ferret samples in GSE147507.  
}
\usage{data(pDf)}
\format{
  A data frame with the following columns.
  \describe{
    \item{\code{id}}{GEO GSM id.}
    \item{\code{cell}}{The cell line or cell type the mRNA was measured in. }
    \item{\code{virus}}{Name of virus infecting the cells. }
    \item{\code{day}}{Time in days after infecting the cells. }
    \item{\code{ser}}{Series of samples to which the sample belongs, numbered 1 to 16. }
  }
}



\keyword{datasets}
