\name{pDh}
\alias{pDh}
\docType{data}
\title{Pheno Data Human (pDh)}
\description{
  Information about the human samples in GSE147507.  
}
\usage{data(pDh)}
\format{
  A data frame with the following columns.
  \describe{
    \item{\code{id}}{GEO GSM id.}
    \item{\code{cell}}{The cell line or cell type the mRNA was measured in. }
    \item{\code{virus}}{Name of virus infecting the cells. }
    \item{\code{hour}}{Time in hours after infecting the cells. }
    \item{\code{ser}}{Series of samples to which the sample belongs, numbered 1 to 16. }
  }
}



\keyword{datasets}
