% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sbr.R
\name{predict.sbr}
\alias{predict.sbr}
\title{Predict SBR/SSBR}
\usage{
\method{predict}{sbr}(object, newdata, coef = "sbr")
}
\arguments{
\item{object}{an object of class 'sbr'.}

\item{newdata}{a (standardized) data matrix from which to predict.}

\item{coef}{choose whether to use the SBR beta estimates (default option "sbr") or the SSBR beta estimates (option "ssbr").}
}
\value{
Returns a vector of predictions.
}
\description{
Predict S3 method for objects of class 'sbr'.
}
\examples{
y <- rnorm(100)
X <- matrix(rnorm(100*300), 100, 300)
G <- gram(X)
model <- sbr(y = y, X = X, G = G, sparsify = TRUE)
predict1 <- predict(model, X[1:10, ], coef = 'sbr')
predict2 <- predict(model, X[1:10, ], coef = 'ssbr')
}
\references{
Perrakis, K. and Mukherjee, S. (2018) Scalable Bayesian regression in high dimensions with multiple data sources. \url{https://arxiv.org/pdf/1710.00596.pdf}
}
\seealso{
\code{\link{predict}}
}
\author{
Konstanstinos Perrakis \email{konstantinos.perrakis@dzne.de}

Sach Mukherjee \email{sach.mukherjee.dzne.de}
}
