% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{launch shiny app}
\usage{
run_app(app, launchBrowser = T)
}
\arguments{
\item{app}{character string with the name of the app.}

\item{launchBrowser}{whether to launch the application with the browser or not.}
}
\description{
launch shiny app
}
\examples{
run_app('dataBrowser')

}
