% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateImg.R
\name{templateImg}
\alias{templateImg}
\title{create an html template for image}
\usage{
templateImg(img, width, height, rainFrom, rainTo, temperatureMax,
  temperatureMin)
}
\description{
create an html template for image
}
