% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templatePastWtr.R
\name{templatePastWtr}
\alias{templatePastWtr}
\title{create an html template for the past weather}
\usage{
templatePastWtr(rain, temperatureMax, temperatureMin)
}
\description{
create an html template for the past weather
}
