% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_balance_by_station.R
\name{water_balance_by_station}
\alias{water_balance_by_station}
\title{Computes the water balance for a certain station (by id or lat long) and saves it in a data.frame.
both data form the beratungsring or the province of Bozen can be used. With mergeBoth = TRUE both are used and averaged.}
\usage{
water_balance_by_station(long = NULL, lat = NULL, idSBR = NULL,
  datestart, dateend, round = "day", provSensor, sbrSensor,
  mergeBoth = FALSE, password, user, host, crop = "tall", taw = 50,
  lmitWarning = 0.8, p = 0.5, startwb = NULL, irrig = 50,
  trgtM = NULL)
}
\arguments{
\item{long}{long input forgetClosestStations}

\item{lat}{lat input for getClosestStations}

\item{idSBR}{id of the SBR station input for getClosestStations}

\item{datestart}{starting date for download}

\item{dateend}{ending date for download}

\item{round}{rounding for resample. Default "day"}

\item{provSensor}{province sencors to dowload}

\item{sbrSensor}{if only data from the province of bozen is supplied se to NULL}

\item{mergeBoth}{if this is true all the measurements from SBR and province are downloaded and the mean is computed. Default FALSE. If true provSensor and sbrSensor are ingnored.}

\item{password}{pasword for beratungsring database}

\item{user}{user for beratungsring database}

\item{host}{host for beratungsring database}

\item{crop}{to compute ET. "short" or "tall". Default "tall"}

\item{taw}{the total available water. Default 50}

\item{lmitWarning}{fraction of the taw below which the app suggests to irrigate. Default 0.8}

\item{p}{fraction of the taw defining the raw (readly available water)}

\item{startwb}{is the initial water balance. Default value is NULL which yields taw*p}

\item{irrig}{vector with irrigation. Same lenght as rows of data or one number for the starting irrigation (first day of the time series)}

\item{trgtM}{a date. if not NULL (default) retains data later than that date}
}
\description{
Computes the water balance for a certain station (by id or lat long) and saves it in a data.frame.
both data form the beratungsring or the province of Bozen can be used. With mergeBoth = TRUE both are used and averaged.
}
