% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB.R
\name{WB}
\alias{WB}
\title{Computes computes water balace from evapotranspiration and precipitation data coming from et function
and getForecast function}
\usage{
WB(data, taw = 50, lmitWarning = 0.7, p = 0.5, startwb = 10,
  irrig = 0)
}
\description{
Computes computes water balace from evapotranspiration and precipitation data coming from et function
and getForecast function
}
