% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge data of a location - using closest province of Bozen and SBR stations -}
\usage{
mergeData(long = NULL, lat = NULL, idSBR = NULL,
  datestart = Sys.Date() - 2, dateend = Sys.Date() + 1,
  round = "day", provSensor = c("GS", "N"),
  sbrSensor = c("Temperatur 2m_min", "Temperatur 2m_max",
  "Relative Luftfeuchtigkeit_min", "Relative Luftfeuchtigkeit_max",
  "Windgeschwindigkeit_avg"), password, user, host)
}
\arguments{
\item{sbrSensor}{if only data from the province of bozen is supplied se to NULL}
}
\description{
Merge data of a location - using closest province of Bozen and SBR stations -
}
