% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeEtAndForecast.R
\name{mergeEtAndForecast}
\alias{mergeEtAndForecast}
\title{merge data from wheater forecast and et average data}
\usage{
mergeEtAndForecast(long, lat)
}
\description{
merge data from wheater forecast and et average data
}
