% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOldAndForecast.R
\name{mergeOldAndForecast}
\alias{mergeOldAndForecast}
\title{merge et and precipitation data from wheater forecast and dayly measured et data}
\usage{
mergeOldAndForecast(data, long, lat)
}
\description{
merge et and precipitation data from wheater forecast and dayly measured et data
}
